release 0.1.120 source code

+++ services/kvs/v1/model/model_scan_kv_response.go
@@ -11,7 +11,7 @@ import (
 type ScanKvResponse struct {
 
 	// 返回的文档数量，为0不表示结束。 > 如果filtered_count超过500仍无匹配，则返回0。
-	ReturnedCount *int32 `bson:"returned_count,omitempty"`
+	ReturnedCount *int64 `bson:"returned_count,omitempty"`
 
 	// 被过滤掉的文档数量。
 	FilteredCount *int32 `bson:"filtered_count,omitempty"`
